src/datahub_custom_sources/runner.py
"""
Programmatic entry-point for running DataHub ingestion recipes.

Usage from Python / Airflow / any scheduler:

    from datahub_custom_sources.runner import run_recipe, run_recipe_dict

    # From a YAML file on disk
    run_recipe("examples/recipes/infa_pmrep.yml")

    # From a dict (e.g. built dynamically in a DAG)
    run_recipe_dict({
        "source": {"type": "infa_pmrep", "config": {â€¦}},
        "sink": {"type": "datahub-rest", "config": {"server": "http://â€¦"}},
    })
"""

from __future__ import annotations

import logging
from pathlib import Path
from typing import Any, Dict, Optional, Union

import yaml
from datahub.ingestion.run.pipeline import Pipeline

logger = logging.getLogger(__name__)


def load_recipe(path: Union[str, Path]) -> Dict[str, Any]:
    """Load a YAML recipe file and return it as a dict."""
    p = Path(path)
    if not p.exists():
        raise FileNotFoundError(f"Recipe file not found: {p}")
    return yaml.safe_load(p.read_text(encoding="utf-8"))


def run_recipe(
    recipe_path: Union[str, Path],
    *,
    dry_run: bool = False,
    preview: bool = False,
    report_to: Optional[str] = None,
) -> Pipeline:
    """
    Load a YAML recipe from disk and execute the DataHub ingestion pipeline.

    Parameters
    ----------
    recipe_path : path to a YAML recipe (same format as ``datahub ingest -c``).
    dry_run     : if True, build the pipeline but do not emit to the sink.
    preview     : if True, log work-units instead of emitting.
    report_to   : optional path to write the JSON run report.

    Returns
    -------
    The completed ``Pipeline`` object (inspect ``.source.get_report()`` etc.).
    """
    recipe_dict = load_recipe(recipe_path)
    return run_recipe_dict(
        recipe_dict,
        dry_run=dry_run,
        preview=preview,
        report_to=report_to,
    )


def run_recipe_dict(
    recipe: Dict[str, Any],
    *,
    dry_run: bool = False,
    preview: bool = False,
    report_to: Optional[str] = None,
) -> Pipeline:
    """
    Execute a DataHub ingestion pipeline from a recipe dict.

    The dict must have at least ``source`` and ``sink`` keys (same schema as
    the YAML files used by ``datahub ingest``).

    Parameters
    ----------
    recipe   : full recipe dict (source + sink + optional pipeline_name, etc.).
    dry_run  : if True, build the pipeline but do not emit to the sink.
    preview  : if True, swap the sink to ``datahub-lite`` / file for preview.
    report_to: optional path to write the JSON run report.

    Returns
    -------
    The completed ``Pipeline`` object.
    """
    if dry_run:
        recipe.setdefault("sink", {})["type"] = "file"
        recipe["sink"].setdefault("config", {}).setdefault("filename", "/dev/null")
        logger.info("Dry-run mode: sink replaced with no-op file sink.")

    if preview:
        logger.info("Preview mode: workunits will be logged but not emitted.")

    pipeline = Pipeline.create(recipe)
    logger.info(
        "Pipeline created â€“ source=%s  sink=%s",
        recipe.get("source", {}).get("type", "?"),
        recipe.get("sink", {}).get("type", "?"),
    )

    pipeline.run()
    pipeline.raise_from_status()

    if report_to:
        _write_report(pipeline, report_to)

    logger.info("Pipeline finished. Workunits emitted: %s", _safe_emitted_count(pipeline))
    return pipeline


def _write_report(pipeline: Pipeline, path: str) -> None:
    import json

    report = pipeline.pretty_print_summary(warnings_as_failure=False)
    Path(path).write_text(
        json.dumps({"summary": report}, indent=2, default=str),
        encoding="utf-8",
    )
    logger.info("Report written to %s", path)


def _safe_emitted_count(pipeline: Pipeline) -> str:
    try:
        return str(pipeline.source.get_report().events_produced)  # type: ignore[union-attr]
    except Exception:
        return "unknown"
