pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "datahub-custom-sources"
version = "0.1.0"
description = "Custom DataHub ingestion source plugins for Informatica, AutoSys, Essbase, SSIS, Ab Initio, and Oracle operational lineage"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Apache-2.0"}
authors = [{name = "Your Team"}]
dependencies = [
  "acryl-datahub>=1.2.0",        # DataHub python + ingestion framework (pydantic v2)
  "pydantic>=2.0",
  "typer>=0.12",
  "rich>=13.0",
  "python-dateutil>=2.9",
  "oracledb>=2.0",
  "requests>=2.31",
  "sqlglot>=25.0",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "ruff>=0.6", "mypy>=1.8"]
airflow = ["apache-airflow>=3.0"]
docs = ["mkdocs-material>=9.0", "mkdocs-awesome-pages-plugin>=2.9"]

[project.scripts]
dhcs = "datahub_custom_sources.cli:app"

# DataHub ingestion plugin entry-points (for `datahub ingest` / `datahub check plugins`)
[project.entry-points."datahub.ingestion.source.plugins"]
infa_pmrep = "datahub_custom_sources.sources.informatica:InformaticaPmrepSource"
autosys_jil = "datahub_custom_sources.sources.autosys:AutoSysJilSource"
oracle_operational = "datahub_custom_sources.sources.oracle_operational:OracleOperationalSource"
essbase = "datahub_custom_sources.sources.essbase:EssbaseSource"
ssis_dtsx = "datahub_custom_sources.sources.ssis:SsisDtsxSource"
abinitio = "datahub_custom_sources.sources.abinitio:AbInitioSource"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"
